✨ Update home dashboard & form rules...

+++ src/lib/services/library.service.ts
@@ -1,9 +1,10 @@
 import { BookcaseModel } from "@/lib/models/bookcase.model";
 import {
   LibraryDocument,
-  LocationModel,
   Location,
+  LocationModel,
 } from "@/lib/models/library.model";
+import { BookModel } from "../models/book.model";
 
 class LibraryService {
   async get() {
@@ -19,12 +20,30 @@ class LibraryService {
   }
 
   async create(payload: Location) {
+    await LocationModel.collection.createIndex(
+      { name: 1 },
+      { collation: { locale: "en", strength: 2 } }
+    );
+
+    const nameCheck = await LocationModel.findOne(
+      { name: payload.name },
+      { collation: { locale: "en", strength: 2 } }
+    );
+
+    if (nameCheck) {
+      throw new Error("Tên thư viện đã tồn tại!");
+    }
+
     const bookcase = await LocationModel.create(payload);
     return JSON.parse(JSON.stringify(bookcase));
   }
 
   async deleteById(_id: string): Promise<void> {
     try {
+      const isLocationClear = await BookModel.countDocuments({ library: _id });
+      if (isLocationClear > 0) {
+        throw Error("Vui lòng chuyển hết sách khỏi thư viện.");
+      }
       await LocationModel.findByIdAndDelete(_id);
     } catch (error: any) {
       throw new Error(error.message);
@@ -45,6 +64,20 @@ class LibraryService {
     updateData: Partial<LibraryDocument>
   ): Promise<LibraryDocument | null> {
     try {
+      await LocationModel.collection.createIndex(
+        { name: 1 },
+        { collation: { locale: "en", strength: 2 } }
+      );
+
+      const nameCheck = await LocationModel.findOne(
+        { name: updateData.name },
+        { collation: { locale: "en", strength: 2 } }
+      );
+
+      if (nameCheck?._id.toString() !== _id) {
+        throw new Error("Tên thư viện đã tồn tại!");
+      }
+
       const data = await LocationModel.findByIdAndUpdate(_id, updateData, {
         new: true,
       });
